feat: add po as realm admin after realm is applied
