+++ app/.env.example
@@ -39,3 +39,4 @@ TF_GH_REPO=
 TF_MODULE_GH_REF=
 GH_ACCESS_TOKEN=
 GH_API_TOKEN=
+IDIR_REQUESTOR_USER_GUID=
