+++ helm/webapp/templates/deployment.yaml
@@ -56,6 +56,7 @@ spec:
             export GH_ACCESS_TOKEN="{{ .Data.data.GH_ACCESS_TOKEN }}"
             export GH_API_TOKEN="{{ .Data.data.GH_API_TOKEN }}"
             export DATABASE_URL="postgresql://${PGUSER}:${PGPASSWORD}@${PGHOST}:${PGPORT}/${PGDATABASE}?schema=public"
+            export IDIR_REQUESTOR_USER_GUID="{{ .Data.data.IDIR_REQUESTOR_USER_GUID }}"
           {{- end }}`}}
     spec:
       initContainers:
